内容回收站调整为逻辑删除实现。 自定义表单调整。

+++ ruoyi-modules/ruoyi-system/src/main/java/com/ruoyi/system/config/I18nMessageSource.java
@@ -33,7 +33,7 @@ public class I18nMessageSource extends AbstractMessageSource implements Initiali
 
 	private Charset encoding = StandardCharsets.UTF_8;
 
-	private int cacheSeconds;
+	private long cacheSeconds;
 
 	private Locale defaultLocale = Locale.SIMPLIFIED_CHINESE;
